+++ src/main/java/ExperienceGroup/Ludora/features/sale/domain/dto/SaleDTOResponse.java
@@ -5,6 +5,7 @@ import ExperienceGroup.Ludora.features.sale.ESaleStatus;
 import ExperienceGroup.Ludora.features.game.domain.dto.GameDTOResponse;
 import ExperienceGroup.Ludora.features.client.domain.dto.ClientDTOResponse;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -12,7 +13,7 @@ import java.util.UUID;
 public record SaleDTOResponse ( UUID externalId,
                                 LocalDateTime date,
                                 ESaleStatus status,
-                                Double totalPrice,
+                                BigDecimal totalPrice,
                                 ClientDTOResponse client,
                                 List<GameDTOResponse> games
 ){
